+++ src/TodoStorage.js
@@ -1,0 +1,14 @@
+const STORAGE_KEY = 'todos-vuejs-demo'
+export const todoStorage = {
+  fetch: function() {
+    let todos = JSON.parse(localStorage.getItem(STORAGE_KEY) || '[]')
+    todos.forEach(function(todo, index) {
+      todo.id = index
+    })
+    todoStorage.uid = todos.length
+    return todos
+  },
+  save: function(todos) {
+    localStorage.setItem(STORAGE_KEY, JSON.stringify(todos))
+  }
+}
